+++ make/source.mk
@@ -149,8 +149,9 @@ COMMON_SRC = \
             io/rcdevice.c \
             io/gps.c \
             io/ledstrip.c \
-            io/osd.c \
             io/pidaudio.c \
+            osd/osd.c \
+            osd/osd_elements.c \
             sensors/barometer.c \
             sensors/rangefinder.c \
             telemetry/telemetry.c \
@@ -289,7 +290,6 @@ SIZE_OPTIMISED_SRC := $(SIZE_OPTIMISED_SRC) \
             config/config_streamer.c \
             i2c_bst.c \
             io/dashboard.c \
-            io/osd.c \
             io/serial.c \
             io/serial_4way.c \
             io/serial_4way_avrootloader.c \
@@ -317,6 +317,8 @@ SIZE_OPTIMISED_SRC := $(SIZE_OPTIMISED_SRC) \
             io/vtx_tramp.c \
             io/vtx_control.c \
             io/spektrum_vtx_control.c \
+            osd/osd.c \
+            osd/osd_elements.c \
             pg/pg.h
 
 # F4 and F7 optimizations
