+++ src/main/io/vtx_control.c
@@ -37,10 +37,11 @@
 #include "fc/config.h"
 #include "fc/runtime_config.h"
 
-#include "io/osd.h"
 #include "io/spektrum_vtx_control.h"
 #include "io/vtx.h"
 #include "io/vtx_control.h"
+
+#include "osd/osd.h"
 
 #include "pg/pg.h"
 #include "pg/pg_ids.h"
