+++ src/test/unit/cli_unittest.cc
@@ -40,11 +40,11 @@ extern "C" {
     #include "flight/servos.h"
     #include "io/beeper.h"
     #include "io/ledstrip.h"
-    #include "io/osd.h"
     #include "io/serial.h"
     #include "io/vtx.h"
     #include "msp/msp.h"
     #include "msp/msp_box.h"
+    #include "osd/osd.h"
     #include "pg/pg.h"
     #include "pg/pg_ids.h"
     #include "pg/beeper.h"
